+++ eslint.config.mjs
@@ -1,0 +1,53 @@
+import globals from "globals";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import js from "@eslint/js";
+import { FlatCompat } from "@eslint/eslintrc";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const compat = new FlatCompat({
+    baseDirectory: __dirname,
+    recommendedConfig: js.configs.recommended,
+    allConfig: js.configs.all
+});
+
+export default [...compat.extends("eslint:recommended", "plugin:react/recommended"), {
+    languageOptions: {
+        globals: {
+            ...globals.node,
+            Atomics: "readonly",
+            SharedArrayBuffer: "readonly",
+            AP: "readonly",
+            window: true
+        },
+
+        ecmaVersion: 2020,
+        sourceType: "module",
+
+        parserOptions: {
+            ecmaFeatures: {
+                jsx: true,
+            },
+        },
+    },
+
+    settings: {
+        react: {
+            version: "detect",
+        },
+    },
+
+    files : [ '**/*.{js,jsx,mjs,cjs,ts,tsx}' ] , 
+
+    rules: {
+        "no-console": "off",
+
+        "no-unused-vars": ["error", {
+            argsIgnorePattern: "^_",
+            varsIgnorePattern: "^_",
+        }],
+
+        "react/prop-types": "off",
+    },
+}];